+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Please also have a look at our
 
 ### Fixed
 
+- Allow very deeply-nested HTML (#1556)
+
 ## 8.2.0: Add support for Symfony 8.0
 
 ### Added